improve bugs.md to allow easier links to a section
As the `bugs.md` has grown, the need to link into a specific
section of the `bugs.md` has become apparent.  While there
are links the entry's `bugs.md` section, due to the
complexity of a head such as:

the links within the header will make it hard for
markdown to html to form such links.

Instead the level 2 header has been transformed into:

vvvvvvvvvvvvvvvvvvv--- level 2 header below ---vvvvvvvvvvvvvvvvvvv

^^^^^^^^^^^^^^^^^^^--- level 2 header above ---^^^^^^^^^^^^^^^^^^^

We made a test modification to `1984/laman/README.md` to
refer to this section as:

vvvvvvvvvvvvvvvvvvv--- reference to the level 2 header below ---vvvvvvvvvvvvvvvvvvv

The current status of this entry is:

```
STATUS: INABIAF - please **DO NOT** fix
```

For more detailed information see [bugs.md#1984/laman](/bugs.md#1984-laman).
^^^^^^^^^^^^^^^^^^^--- reference to the level 2 header above ---^^^^^^^^^^^^^^^^^^^

One advantage of the above section is that it contains only
a single link into `bugs.md`.

In particular we use the following string as a link into `bugs.md`:

    /bugs.md#1984-laman

We plan to verify that this is the proper way to link into `bugs.md`
after we push this commit.

During these edits, we discovered a set of quasi-duplicated
big entries spanning from 1998 thru 2013.  We say quasi-duplicated
because the repeated information was not the same.

We made a guess in `bugs.md` as to which parts were correct,
however our guess may have been incorrect.  Therefore, we
preserved the two sets in two files.

The 1st set of 1998 thru 2013 is found in: `bugs-1998-2003-parta.md`.
The 2nd set of 1998 thru 2013 is found in: `bugs-1998-2003-partb.md`.

Please inspect the sets, and if needed, update `bugs.md` as needed.
Please remove `bugs-1998-2003-parta.md` and `bugs-1998-2003-partb.md`
when you are done.