Update 1991/westley in bugs.md
Added to the INABIAF that not only is there a way to cheat which was
documented at the time and which I discovered when fixing the code but
there is actually a way that I discovered recently where you can cheat
but it still actually wins! Does that mean it is cheating as well? :-)
Or is it that one cannot actually cheat a cheater? :-) Either way this
felt like a good thing to add to the bugs.md as well.