Fixes in 2000/primenum - int main from void main
To make the entry more portable the void main was changed to int main.

The no arg check was NOT fixed: this is a documented feature not a bug
and should not be fixed. If one wants to fix it for personal enjoyment
(if they can get pleasure from a couple seconds :-) ) they may do so but
it is not to be committed.