Minor update 2001/bellard in bugs.md
Fix a few things. Now it's in good state. There were some fixes made
already but it still only works for i386 linux (except the supplementary
program can compile and work in modern systems for fun after I fixed it)
and if anyone wants to provide a version for x86_64 linux, for example,
they can do so.