Minor updates to 1984/mullender
Minor corrections in the README.md and bugs.md.

The README.md was not checked for other issues at this time. When that
happens 1984 should be complete in every way except that 1984/README.md
has to be checked again but that will happen after the other entries of
all the other years are done.