Add tmp/format-headers.sh script from Landon
With a slight modification so that the script uses git ls-files so that
it only acts on files called README.md under the directories of the
years (i.e.  '[0-9]{4}/*README.md') that are under git control this
script formats headers so that the spacing before lines starting with
'##' (as in '^##') are correct. It checks for specific header names.
More could be added if necessary and if any heading level is wrong or it
is decided that a different heading level might be better the regexp can
be modified to fix it as well.

This script will be useful to run after every README.md file has been
updated.

Added to tmp/ as it is likely (or possibly?) a temporary tool that might
not be needed after the temp-test-ioccc repo is merged back into the
winner repo. On the other hand it might be useful to use after new
contests are run or even after some major edits of README.md files.

Uses perl in a bash loop.