Add 1998/schweikh3/try.sh + other fixes & updates
The try.sh script runs the commands in the try section of the README.md.

The README.md had SOME format fixes.

The Makefile symlinks the program to samefile.

The author stated on their website that:

    the Internet Systems Consortium (ISC) has added my samefile utility
    to its treasure of software gems.

and this has been added to the README.md in the judges' remarks, along
with the web page for the program!