Remove outdated feature/bug in 1991/dds/README.md
As the definition of a bug and feature was changed numerous times the
feature or bug of if a file could not be opened for reading or a file
could not be opened for writing was actually fixed to not crash. This
has not been done in other entries and almost certainly won't be but
with this one it was done (recently when there was a change of bug
status in several entries as how the bugs.md file being referred to in
the README.md files was improved) so this has inconvenient feature was
removed from the README.md file. It was not noted in the thanks file
either so that has been updated as well.