Fix 1990/jaw/btoa script
This now lets the command in the judges' remarks:

    echo "Quartz glyph jocks vend, fix, BMW." | compress | ./btoa | ./jaw

not report just:

    oops
    oops

and instead it outputs data which suggests, I believe, that the program
is not buggy but instead the script had to have the 'xbtoa' changed to
'btoa'.

Updated bugs.md and thanks-for-fixes.md and README.md file too.

Other changes were made to thanks file as they are more appropriate than
in the README.md (of 1990/theorem) after the (once again sounding like
a broken record) vastly superior idea of the thanks file came up. This
addition here is unintentional; it was just already there but not ready
for committing as 1990/theorem/README.md was not yet ready. That file
will be committed today.