Don't hard code paths in 1990/jaw/shark.sh
It is not actually necessary to mess with the path and in fact the $PATH
variable is no longer modified or even referred to. The purpose of the
program was that even without btoa and uncompress the script should be
able to extract the tarball.

Now only the tarball path is hard coded as reading a tarball from stdin
is problematic though I'm pretty sure it used to work.