Minor updates to 1991/fine
I updated the try.sh with a new string that I discovered that is for fun
output especially given something the judges say in the guidelines.

Also since some of the strings are from Brian Westley's entry from 1989
and another one was credited to him I briefly referred to the entry as
well.