Typo fixes in index.html
Including one that has bugged me for years now, the 'Ideas of March' :-)
instead of the Ides of March.

Fun facts: ides can be used for other months too, in Roman times. It was
the middle of the month so depending on the month it would be the 15th
or the 13th. 'Calends'/'kalends' for first day of the month and 'nones'
for before they had months! :-) (no: actually it was the ninth day
before the ides by inclusive reckoning so the 7th or 5th depending on
the month).