Further clang fixes for 1991/dds
It came to my attention that some versions of clang trigger and
additional warning that has to be disabled in order to compile the a.c
that the entry generates. This has been fixed.

I am NOT including the translate.c file that I wrote to come up with the
proper string (it encrypts and decrypts though with some limitations) as
I don't want to give it away too much. I MIGHT include it on my website
another time in case it is ever necessary (though the hope and belief is
it won't be but who knows with clang). This would be AFTER the website
has been looked at. This has already taken too much time to fix. It is
now time to finish 1992 and then go on to 1993.