Format/typo check + update bugs.md for 1992/vern
The bugs.md has both a feature and a bug or maybe mis-feature that might
be good to fix. I gave a tip in the bugs.md that might help with parsing
the code a bit.

The README.md file has been typo and format checked/fixed.

This should complete 1992.