Update FAQ wrt X11 entries in macOS
There are three new images in png/ which are part of this entry. It
might be reasonable to remove ne but to keep it consistent with the rest
of the answer it is kept there for now.

This commit is partly a test to see how it comes out with GitHub and
also in general so that it might have to be revised.

The faq.md has had some html added as well, not only embedded CSS style
for 'img' to hopefully make it responsive but also the question has a
named anchor to link to more easily later on (this is likely to happen
to the rest of the questions/answers at another time).