Fix bug of 1994/dodsond2 wrt arrows stolen
The problem was that one could never get their arrows back from the
robber if they ended up shooting an arrow because that means the number
of arrows stolen could never be the right number of arrows. Now if the
number of arrows stolen (when you encounter the robber) is equal to the
number of arrows you found - the number of arrows you shot you get
arrows back: as long as the robber took some of course. At this point
the counters are reset so that the check does not fail the next time.
Previously it compared to just 3 as there are only three arrows. But in
the alt version (which was helpful to test this out more) this is
obviously a problem even if it wasn't in the original. It was however a
problem in the original because if you shot an arrow you could not get
them back even if you have zero and encounter the robber. Of course now
the check has to verify first (&&...) that the number of arrows stolen >
0 as otherwise the variables could get messed up.

This should actually complete 1994/dodsond2 unless of course I missed
something/made a mistake here.