Add 1996/schweikh2/try.sh + format/typo checks
The try.sh script runs some of the commands in the try section as well
as a minimal version of something the author recommends one trying for a
pleasing effect. The reason it is minimal is because it is an infinite
loop and it seems like not a good idea to have an infinite loop in a
script that demonstrates a few things of the entry.

Format and typo check the README.md file.