Make 1992/kivinen/kivinen.c alt code
Originally the main version was changed to have a call to usleep(3) as
the game moves so fast. It was discussed a while back however to in
cases like this to make it so that the build/use/try sections refer to
the alt code and then instead of an alternate code section have an
original code section. Because of this the modified code was changed to
be an alt version and then restore the original with the fixed code.

As for the fixed code (and also alt code) I have made it look more like
the original by changing the 'ext' macro to be 'exit' (with the comma
operator). I tried and believe I succeeded (mostly if not totally) in
making the lines start and end on the same columns and in at least some
lines the same character starts and ends the lines as the original.