Updates and fixes to 1998/schweikh1
The perl script that the author said they use to count chars in the code
(according to the IOCCC rules at the time) was added as charcount.pl.

Moved the macOS details/spoiler (though it has information about the
main entry as well) to macos.md.

README.md file format and typo checked.

This should complete this entry's review.