Updates and fixes to 1998/fanf
The try.sh script has been added to run the various inputs the judges
recommended one try.

The Makefile now can do the intermediate steps. This is possible with
the way I fixed the entry to compile. If you look at the original code
it's different than the fixed version which is post processing (if I
recall correctly).

Updated .gitignore.

Unless there's a mistake in markdown formatting this should complete
1998/fanf's review.