Add 1998/dloweneil alt code + formatting
The alt code should feel more at home for vi(m) users with movements.
With the exception of the drop key (d in the original) you can use the
original keys as well. Keys to drop the letter in alt version j or
space, whichever you prefer. Also you can quit the game gracefully (q).
That was something the author(s)? noted as something they wish they
could have done but due to contest size restrictions they could not.

The README.md file has been format and typo checked and this should
complete the review of this entry.