update FAQ and add FAQ section numbers
A number of the FAQ questions strings were rewritten to be order,
a big more consistent in tone, and without any markdown symbols.

Each FAQ section has an HTML name tag before the section.
This has been done as an experiment in linking.

The FAQ table of contents has links into the FAQ sections,
however those links are using the current HTML name tags
and may not work in the `faq.md` markdown document.
We are not sure this will work, but we followed this suggestion:
https://stackoverflow.com/questions/6695439/how-to-link-to-a-named-anchor-in-multimarkdown