Add 1998/bas2/try.sh + format/typo check README.md
The try.sh allows one to pass in args: if a readable regular file it
will be fed to the program and otherwise it'll act as if it's a string.
It also has some additional commands that I thought might be fun to show
the entry.

Format/typo check README.md file.