Improve try.sh and compile.sh in 1989/westley
The scripts now allow one to pass in CC=foo. If a path is specified
it'll use that path (for instance in case you want a different version
or because in macOS /usr/bin/gcc is actually clang); otherwise the shell
will try working it out. If one does not do CC=foo then it'll default to
'cc'.

Both try.sh and compile.sh allow this and try.sh will pass CC to
compile.sh before running the commands.

The CSILENCE variable in the Makefile has been updated as well. It is
not possible to silence all warnings with clang but most of them are, at
least in macOS clang (not tested linux yet).