Make 2013/birken restore sanity after exit
As long as the program (alt code and otherwise) runs to completion the
program will return instead of 0:

    system("reset");

so that the terminal becomes sane again (otherwise cursor is invisible).
One could also use 'tput cvvis' but this is longer and I wanted to make
the column as close to the original ending as possible.