Fix infinite loop in 2000/robison w/invalid input
If someone entered invalid input (say because they don't know the format
yet or because they make a typo or some other reason) their terminal
would be flooded with an error message, forcing them to kill the game.
This was done in much the same way as the fix for 1987/lievaart. Without
this fix the words:

    Black position and direction: illegal

would flood the screen.

The Makefile CINCLUDE now has -include stdio.h (not strictly necessary
however).

Format and typo check README.md.