Update bug status of 2000/dlowe
This entry crashes in macOS. This MIGHT be related to the version of
perl, both with MacPorts and Homebrew. See the bugs.md for many more
details.

The author gave an example command and output (well several) and one I
tried (I did not try all) did not report correct output (some did).
Again the bugs.md file has many more details.

Format/typo fix README.md.

With the exception of possibly fixing the bugs noted in bugs.md this
should complete the review of 2000/dlowe.