Fix fix on 1991/buzzard
The parameter type of q in function z() should be a 'char *', not a
'char **'.

Added to CSILENCE -Wno-implicit-int.

Note that if including files for abs(3), strlen(3) or other functions
implicitly declared the program will not work (or at least it won't work
if at least one of them is #included?).