Add 2000/bmeyer/try.sh + formatting/typo checks
The try.sh script has improved the judges' suggested commands in that it
does not assume a certain number of columns. Instead it determines the
number of columns and uses that value. It warnings (before starting)
that if it's <81 it might not show up well. It also warns that it should
be white text on a black background.

The Makefile copies the program to the name the author gave the program.

The README.md file has been format/typo checked and fixed.

This completes the review of 2000/bmeyer and it should also complete the
review of 2000 the 2000/README.md file notwithstanding (this goes for
all the YYYY/README.md files).