Fix 2001/ollinger bug status in bugs.md
There was some inconsistency as far as whether or not the problem was a
bug or feature, where not enough args would result in a crash. At one
point it seems it was changed to a bug, not a feature, and it is then
that I must have fixed it. However the bugs.md file incorrectly labelled
another entry as ollinger and the README.md for ollinger referred to
something that did not exist or rather it existed but it was for another
entry.

A minor fix in bugs.md has been done as well: change the status name of
using fgets() to not refer to getline(3); rather the description does
that. This is because no other place has that function in the status
line. Doing this is how I found the problem this commit fixes as another
entry should have that status.