Fixes to 2001/westley
The Makefile now always regenerates the .sort.c and the .punch.c
versions if westley.c has been modified.

The try.sh script had a typo (not in a command that ends up being run
but rather showing what would be run).

Add bugs and (mis)features section with two statuses.

Format/typo check README.md file.

It is not clear to me just yet if this completes the review of the entry
but it might.