Add bug status to 2001/williams
There are two bugs that are known. First is that at the end of the game
(when you lose) it appears to in (an infinite?) while loop call
usleep(3). However you can still shoot your missiles at this point. The
other bug is that it will eventually crash. It's easy to make it crash
if you wait until everything is destroyed (that is when it starts
printing the lines in a loop, sleeping over and over again) and only
then start shooting your missiles, over and over again until it crashes.
It appears that it's the 'Window *d' getting corrupted but I'm not sure
why. These details have been added to bugs.md.

The README.md file has been format fixed.

There was a minor change to williams.c as well though this was only
rolling back a change that should not have been committed yesterday.