Updates to 2001/herrmann1
Update plus1.turing file I added a while back (based on author's
remarks).

One of the missing files was not actually missing; it was a misreading
of the author's remarks. It was suggesting that if one wanted to make
their own file by that name, not that the file was there. There still is
a missing file, however.

Format/typo check README.md.