Add 1998/schweikh3 alt code
This is for the rare case that the author described where the size needs
to be increased: it can now be done when compiling without having to
update the code. The original code is the recommended one try first; if
any of '-D' macros are not defined it will use the default, including
the size one. This so to make it easier to redefine just one, rather
than having to specify all the -D options.