Update CNAME
Due to a request I had (allowing both www.ioccc.org and ioccc.org) a
while back Landon fixed it but this change was not done in this repo.
The hope is that this commit will mean it won't be a problem when this
repo is merged back into the winner repo.