Fix infinite loop in 1992/vern
If one entered input < 0 || > 077 (octal) the program would enter an
infinite loop that flooded the screen, forcing the user to terminate the
program.

The problem was that scanf(3) (with "%o %o" specifier string) was in a
loop checking for return value of 2 (this was not a problem because
scanf(3) does not return until all the specifiers are processed or I
suppose some error occurs) and not checking the values. Instead it just
checked a function on the numbers. This fix involves using scanf(3) with
the specifier "%4s %4s" and then using strtol(3) with base 8 and then
checking for < 0 || > 077 (using "%o %o" will not work).

The game now functions as expected without flooding the terminal if
invalid input is given. The judges' remarks about the range has been
kept in as it's still invalid it just won't flood the screen.