Fixes in 2005/sykes
Add PET program written in the emulator and add it to the try.sh script.

The try.sh script now explicitly requires one hit 'q' (or 'Q') to quit
(or else killing it or hitting interrupt); after one command is done
(hitting ctrl-c) it shows the menu again which has the quit option in
there. This makes it a bit easier or at least clearer (cleaner?).

The Makefile DATA variable had to be updated after file rename.

There were some other issues in the README.md file as well that I missed
yesterday.

The bugs.md file has been updated pointing out that one has to hit
ctrl-c or whatever the interrupt is set to.

I believe now that this entry should be complete.