Update 2006/monge bugs status
There was an inconsistency between the README.md and the bugs.md file
that has been fixed.

As well the program requires one be able to read, write and execute
memory. This is not allowed in macOS with the silicon chip so I have
added the status of it not working with some platforms.

A few minor formatting fixes were made in the README.md file but I
didn't go through the file so it can't be considered done.