Make 2006/sykes2 main() have two args
This is out of an abundance of caution as some of the versions of clang
report that main() can only have 0, 2 or 3 args to main(), not 4 or
anything else. The version (clang version 17.0.4) I have encountered
this with actually allows 1 arg but since clang is so picky I decided to
make a second unused arg (a char **) in case the devs decide to ruin
even more fun programs. There are other entries that also use one arg to
main() and I will try locating those too, first with the entries I have
not reviewed yet and then later on (time constraints notwithstanding)
the ones I have already reviewed.

Makefile updated with new -Wno- options for clang with linux.