Format/typo check 2006/stewart
The .gitignore has been updated to have '*.xbm' and '*.png' as it seems
that the point of maze.xbm and maze.png etc. are to not include the
images generated by the program and this is cleaner and easier.

The README.md file has been format/typo checked.

This should complete the review of 2006/stewart.