Add alt versions to 1991/cdupont
There are three alt versions. One has a comment and an unused goto label
removed that are warned against being removed; another has this done and
is also beautified (which is also warned against) and the third is one
that is just beautified.

This involved changing the 's' macro to be '__FILE__' because otherwise
it would read in the original code and not show the problem.