Minor update to 1994/weisberg
To make the entry a bit more user friendly and also easier to use with
other tools (in a pipeline - if anyone wanted to for some reason) the
Makefile was modified to print a newline after each number rather than a
space.

I noticed that the Alternate code section had no subsections for
'Alternate build' or 'Alternate use' and these were added as well. I am
not sure how this happened but I might do a quick pass through over all
entries once everything else is done just to be sure.