Add alt code to 2011/richards + format/typo checks
The alt code is a starting point for macOS should anyone be brave enough
to try solving the problem of running mmap()d code. This does appear to
be an arm64 thing so anyone trying it will have to have a Mac with the
silicon chips. The bugs.md and thanks file both link to it.

Format/typo check README.md.