Add 2011/hamaji/try.sh, .nono files + formatting
The .nono files are more data files, some based on the authors' remarks
and two from a website they linked to.

The try.sh script runs the program on those files, others that were
already provided and if perl is installed it verifies that the perl
script will generate the same output as the entry being fed dragon.nono.

Format/typo check README.md file.

This should complete the review of 2011/hamaji.