Add demo video to 2011/eastman + format/typo check
The video file is of the Amiga Boing Ball. It is not embedded in the
README.md file as that's not been worked out yet.

The README.md file has been format/typo checked. With the exception of
hopefully getting the video embedded in the README.md this should
complete the review of 2011/eastman.