Add 2011/fredriksson/try.sh + format/typo checks
Before the last command that the try.sh script will do it first tries to
set the stack size to unlimited (for the process) and if it fails it
won't run the last command as it will likely dump core.

Add entry to bugs.md but since there are quite a few things I just
linked to the author's remarks (the sections). This can be adjusted
another time if necessary.