Add make test rule to 2012/zeitak + formatting
The make test rule runs the new test.sh script which checks the return
value of the program run on a variety of new files (and the old
incorrect.c which was renamed incorrect.paren.c). The new files include
1984/anonymous.c and 1984/anonymous.alt.c as the author explicitly
mentioned that entry.

The README.md file has been format/typo checked.