Update 2012/deckmyn - try.sh, bugs.md, formatting
The try.sh script is as usual.

The bugs.md has some features, not bugs.

The README.md file has been format/typo checked.

The deckmyn.md file has been format/typo checked.

A caveat with this entry is I was not able to verify as much when it
comes to checking things as I do not have a musical background (despite
the fact my family is full of musicians).