Typo fixes in news.html
There were references to hint.html which were changed to README.md
files. Later on I presume they'll be index.html files on the main
website but for now these do not exist as such.

The number of changes in the repo include over 2600 now. I kept the word
nearly (next to 2645) because there have been some changes that were
reverted or decided against, for example.

There was a typo in references to GitHub too.