Add alt code to 2012/omoikane + try + formatting
There are two alt versions which should be useful for (respectively)
those who do not have /dev/urandom and those who use Windows. Both of
these versions read from argv[0] instead of /dev/urandom but the second
one sets binary mode on stdin and stdout. This is to be consistent with
the fact that other entries also have this as an alt version.

The try.sh and try.alt.sh scripts work on the entry and the first
version of the alt code, the one that is not Windows specific. The
scripts consist of commands from both the to use section of the
README.md (which is still there) and the try section (which now refers
to try.sh).

Format/typo checks in README.md file.