Swap alt and alt2 of 2012/endoh1
The alt2 is the deobfuscated version; the alt one is the one with the
ability to slow down/speed up the fluid and control the gravity factor,
the pressure factor and/or the viscosity factor.

The alt version no longer imposes a lower limit on the factors because
doing so prevents the ability to define them to 'I' which the author
noted is a functional difference.