Improve 2012/endoh1 alt code, add try.alt.sh
The alt code now has an option to use alarm(3) so that one need not hit
ctrl-c/intr for each run (default 10 seconds, can be disabled or changed
at compilation just like the other macros). This is used in try.alt.sh
which shows the program (colour version) on the source code and then the
text files provided by the author. The script does two 'versions' for
each file: one with setting gravity factor to I (see author's remarks)
and one without that. These changes make a nice demonstration of the
entry.

This commit also fixed a problem that the previous commit caused to do
with the alt targets in the Makefile.

This should complete the review of 2012/endoh1, leaving two more to go
for 2012.