Update 2013/birken/try.sh + format/typo checks
The try.sh script now does a random run of the program (previously in
the remarks in the try section) involving perl but only if perl is
installed in the system.

The README.md file has been format/typo fixed.

It is not clear if this entry is done or not yet as it might be improved
if try.sh is renamed try.alt.sh (for the alt version which it uses) and
a new try.sh be formed which uses the original code so that one does not
have to do say:

    BIRKEN=birken ./try.sh

This probably would be better but it is not done yet.