Remove 2013/cable3 from bugs.md
As the bug status was a missing file and it was actually hd.img and as
that was added to the repo the entry can be removed from the bugs.md
file.

I also added some important points for the entry about quitting
(elaborating) and also how to input certain keys like alt/ctrl
combinations (amongst others) that the author had later on in the
README.md file. The former might especially be important if someone did
it at the console and was not aware of how to open another console but
even so it can be annoying having to close a tab just to stop a program.

Now I can say that 2013/cable3 is almost certainly complete (review
wise?).