Add 2013/endoh1/try.sh + format/typo checks + bugs
The bugs.md file has limitations of the entry added to it.

The README.md file was format/typo checked.

The try.sh script is as usual. I also added a command to try in the
'try' section that is not in the try.sh script due to it being long
output but still fun to try.